refactor(canvas): extract dock tab strip
